requirements.txt
# The basic requriements.
typing-extensions>=4.1.0
pydantic>=2
markdown-it-py[linkify]
beautifulsoup4
pillow
httpx
rapidfuzz
